web_framework/requirements.txt
python-dotenv
flask
boto3
requests
gunicorn